Restructured library packages
