+++ src/main/kotlin/no/nav/emottak/utils/kafka/client/EventPublisherClient.kt
@@ -1,4 +1,4 @@
-package no.nav.emottak.utils.kafka
+package no.nav.emottak.utils.kafka.client
 
 import io.github.nomisRev.kafka.publisher.KafkaPublisher
 import no.nav.emottak.utils.config.Kafka
@@ -6,7 +6,7 @@ import no.nav.emottak.utils.config.toKafkaPublisherSettings
 import org.apache.kafka.clients.producer.ProducerRecord
 import org.apache.kafka.clients.producer.RecordMetadata
 
-class KafkaPublisherClient(
+class EventPublisherClient(
     private val config: Kafka
 ) {
     private val kafkaPublisher = KafkaPublisher(config.toKafkaPublisherSettings())
